rewrite store and fixes
