+++ src/pages/main/components/offers-list.tsx
@@ -2,8 +2,8 @@ import { FC, useCallback, useEffect, useState } from 'react';
 import { Offer } from '../../../types/offer';
 import Card from './offer-card';
 import { useDispatch } from 'react-redux';
-import { changeActiveOfferId } from '../../../state/actions';
 import React from 'react';
+import { changeActiveOfferId } from '../../../state/offer/offer-actions';
 
 export interface OffersListProps {
     offers: Offer[];
