+++ src/state/city/city-actions.ts
@@ -1,0 +1,4 @@
+import { createAction } from '@reduxjs/toolkit';
+import { City } from '../../types/offer';
+
+export const changeCity = createAction<City>('city/changeCity');
